pyproject.toml
[project]
name = "dear-agent"
version = "0.1.0"
description = "DEAR Agent - Digital Enlightenment & Artistic Reasoning"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    # LangChain 1.x (2026 最新)
    "langchain>=1.2.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    # LangGraph
    "langgraph>=0.2.70",
    # 火山引擎
    "volcengine-python-sdk[ark]>=1.0.0",
    "websockets>=14.0",
    "httpx>=0.27.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
    # MinIO 对象存储
    "minio>=7.2.0",
    # 数据库
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "C4"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
